Make sure disk syncs as part of match exactly sync
